Changed from byte to int

+++ code/entities/map/SpectatePoint.cs
@@ -7,7 +7,7 @@ namespace Paintball;
 [Hammer.EntityTool( "Spectate Point", "PaintBall", "Defines a fixed point where a player can spectate from." )]
 public class SpectatePoint : Entity
 {
-	[Property] public byte Index { get; set; }
+	[Property] public int Index { get; set; }
 
 	public override void Spawn()
 	{
